backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
groq==0.4.2
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
twilio==9.3.2
pydub==0.25.1
httpx==0.27.2
